+++ src/ansible_navigator/ui_framework/ui.py
@@ -45,7 +45,7 @@ from .ui_config import UIConfig
 from ..utils import templar
 
 
-from ..yaml import human_dump
+from .._yaml import human_dump
 
 STND_KEYS = {"^f/PgUp": "page up", "^b/PgDn": "page down", "\u2191\u2193": "scroll", "esc": "back"}
 END_KEYS = {":help": "help"}
